// src/data/types.d.ts
interface VenueData {
  konum: string | null;
  isim: string | null;
  kategoriler: string | null;
  priz: string | null;
  wifi: string | null;
  wifiHiz: string | null;
  gurultu: string | null;
  calismaSaatleri: string | null;
  instagram: string | null;
  harita: string | null;
  notlar: string | null;
}

export type { VenueData };